Generalize Validator to check generic components
