+++ src/main/kotlin/org/opensearch/observability/validation/IntegrationComponent.kt
@@ -1,0 +1,5 @@
+package org.opensearch.observability.validation
+
+enum class IntegrationComponent(val resourcePath: String) {
+    INTEGRATION("/schema/system/integration.schema")
+}
