--- src/main/kotlin/org/opensearch/observability/validation/IntegrationValidationException.kt
@@ -1,4 +1,0 @@
-package org.opensearch.observability.validation
-
-class IntegrationValidationException(message: String? = null, cause: Throwable? = null) : Exception(message, cause) {
-}
