+++ src/main/kotlin/org/opensearch/observability/validation/ValidatorException.kt
@@ -1,0 +1,4 @@
+package org.opensearch.observability.validation
+
+class ValidatorException(message: String? = null, cause: Throwable? = null) : RuntimeException(message, cause) {
+}
